# pos/openapi/RmsSupportHub.Pos.Agent.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "RMS+ POS Agent API",
    "description": "Destination-owned foundation contract for the per-device Windows POS Agent. The Agent exposes a fixed HTTPS loopback endpoint for direct Support Hub browser access. Windows Negotiate supplies the authenticated identity, local Built-in Administrators membership authorizes protected reads and typed Windows-service and RMS database operations, and no Support Hub backend relay is involved. The Agent exposes read-only device/configuration/RMS visibility plus allow-listed service control and server-owned RMS database backup/restore workflows. Downloader, maintenance, and artifact capabilities remain typed, principal-scoped, and fail-closed. Mutation tokens are short-lived, one-use, and bound to the authenticated principal, exact Origin, registered operation, target, method, and path.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://rms-pos-agent.localhost:5001",
      "description": "Fixed HTTPS loopback Agent origin"
    }
  ],
  "paths": {
...
    "/api/v1/packages/status": {
      "get": {
        "tags": [
          "Agent Package"
        ],
        "summary": "Read the server-owned Agent package status",
        "description": "Returns only safe installed-manifest metadata and truthful verification state. Package paths, archives, private keys, and installer arguments are never exposed.",
        "operationId": "GetAgentPackageStatus",
        "responses": {
          "200": {
            "description": "The Agent returned the typed AgentPackageStatusDto status projection.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentPackageStatusDto"
                },
                "example": {
                  "verification": "unknown",
                  "state": "notAttempted",
                  "detail": "No verified server-owned package is available."
                }
              }
            }
          },
          "403": {
            "description": "AuthorizationMiddleware may reject a non-Administrator with a bodyless response. If the exact-origin transport gate rejects the browser origin, the safe problem code is origin_rejected.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProblemDetailsDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/packages/preview/{operationId}": {
      "get": {
        "tags": [
          "Agent Package"
        ],
        "summary": "Preview a typed Agent package operation",
        "description": "Reads only the Agent-owned package catalog and verifier. Preview does not stage, activate, uninstall, register, or roll back a package.",
        "operationId": "PreviewAgentPackageOperation",
        "parameters": [
          {
            "name": "operationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The Agent returned the typed AgentPackagePreviewDto preview.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentPackagePreviewDto"
                },
                "example": {
                  "previewId": "opaque-preview-id",
                  "operation": "upgrade",
                  "ready": false,
                  "verification": "unknown",
                  "blockers": [
                    "package_unavailable"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The Agent rejected a non-canonical host with host_rejected or a non-HTTPS request with https_required; the response uses the safe Agent problem-details contract.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "The request host is not accepted.",
                  "status": 400,
                  "code": "host_rejected",
                  "correlationId": "example-correlation-id"
                }
              }
            }
          },
          "403": {
            "description": "AuthorizationMiddleware may reject a non-Administrator with a bodyless response. If the exact-origin transport gate rejects the browser origin, the safe problem code is origin_rejected.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProblemDetailsDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/packages/operations": {
      "post": {
        "tags": [
          "Agent Package"
        ],
        "summary": "Start an authorized typed Agent package operation",
        "description": "Accepts only a bounded typed request after server-owned validation. Destructive or side-effecting dispatch requires the exact-origin, administrator, one-use method/path mutation-token, challenge/confirmation, and idempotency boundaries; ambiguous outcomes are retained and never retried automatically.",
        "operationId": "StartAgentPackageOperation",
        "requestBody": {
          "description": "The browser supplies only logical IDs, exact confirmation where required, and a bounded idempotency key. Server-owned paths, credentials, SQL, and service targets are not accepted.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentPackageOperationRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The Agent accepted or safely rejected the typed package lifecycle request with truthful rollback and recovery state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentPackageOperationDto"
                }
              }
            }
          },
          "400": {
            "description": "The Agent rejected the typed request, challenge, configuration, or bounded idempotency key with safe problem details or typed NotAttempted state.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProblemDetailsDto"
                }
              }
            }
          },
          "401": {
            "description": "The Windows authentication middleware issued a Negotiate challenge for this protected mutation.",
            "headers": {
              "WWW-Authenticate": {
                "description": "Negotiate challenge emitted by the Windows authentication middleware.",
                "schema": {
                  "type": "string",
                  "example": "Negotiate"
                }
              }
            }
          },
          "403": {
            "description": "Authorization, exact-origin transport, SID resolution, or the one-use mutation-token boundary rejected the mutation request.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProblemDetailsDto"
                }
              }
            }
          },
          "500": {
            "description": "The Agent returned a safe generic server-error response without exception, credential, path, or SQL details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProblemDetailsDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/packages/operations/{operationId}": {
      "get": {
        "tags": [
          "Agent Package"
        ],
        "summary": "Read one principal-scoped Agent package operation",
        "description": "Reads principal-scoped, sanitized progress and outcome for one retained Agent operation. Operation IDs and artifact capabilities are opaque; raw paths, credentials, SQL, service names, and exception text never cross the response boundary.",
        "operationId": "GetAgentPackageOperation",
        "parameters": [
          {
            "name": "operationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The Agent returned the retained typed package lifecycle state and recovery evidence.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentPackageOperationDto"
                },
                "example": {
                  "operationId": "opaque-operation-id",
                  "state": "running",
                  "progressPercent": 50
                }
              }
            }
          },
          "404": {
            "description": "The opaque operation identifier is not retained for the authenticated principal."
          },
          "403": {
            "description": "AuthorizationMiddleware may reject a non-Administrator with a bodyless response. If the exact-origin transport gate rejects the browser origin, the safe problem code is origin_rejected.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProblemDetailsDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentPackageFileDto": {
        "required": [
          "logicalName",
          "sizeBytes",
          "sha256",
          "required"
        ],
        "type": "object",
        "properties": {
          "logicalName": {
            "type": "string",
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          },
          "sizeBytes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed.",
            "format": "int64"
          },
          "sha256": {
            "type": "string",
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          },
          "required": {
            "type": "boolean",
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          }
        },
        "description": "Safe logical package-file metadata used for server-side checksum and size verification."
      },
      "AgentPackageManifestDto": {
        "required": [
          "packageId",
          "version",
          "supportedOperatingSystem",
          "supportedRuntime",
          "serviceDisplayName",
          "serviceIdentity",
          "scmName",
          "signatureAlgorithm",
          "signerDisplayName",
          "packageSha256",
          "files",
          "aclRequirements",
          "certificateRequirements",
          "previousVersion",
          "rollbackAvailable"
        ],
        "type": "object",
        "properties": {
          "packageId": {
            "type": "string",
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          },
          "version": {
            "type": "string",
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          },
          "supportedOperatingSystem": {
            "type": "string",
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          },
          "supportedRuntime": {
            "type": "string",
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          },
          "serviceDisplayName": {
            "type": "string",
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          },
          "serviceIdentity": {
            "type": "string",
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          },
          "scmName": {
            "type": "string",
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          },
          "signatureAlgorithm": {
            "type": "string",
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          },
          "signerDisplayName": {
            "type": "string",
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          },
          "packageSha256": {
            "type": "string",
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          },
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentPackageFileDto"
            },
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          },
          "aclRequirements": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          },
          "certificateRequirements": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          },
          "previousVersion": {
            "type": [
              "null",
              "string"
            ],
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          },
          "rollbackAvailable": {
            "type": "boolean",
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          }
        },
        "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
      },
      "AgentPackageOperationDto": {
        "required": [
          "operationId",
          "operation",
          "state",
          "outcome",
          "progressPercent",
          "stage",
          "detail",
          "rollbackAttempted",
          "rollbackSucceeded",
          "recoveryRequired",
          "startedAtUtc",
          "completedAtUtc",
          "correlationId"
        ],
        "type": "object",
        "properties": {
          "operationId": {
            "type": "string",
            "description": "Opaque Agent-issued identifier scoped to the authenticated principal and typed workflow."
          },
          "operation": {
            "description": "Typed package lifecycle operation selected by the server-owned policy.",
            "$ref": "#/components/schemas/AgentPackageOperationKindDto"
          },
          "state": {
            "description": "Typed package lifecycle state.",
            "$ref": "#/components/schemas/AgentPackageOperationStateDto"
          },
          "outcome": {
            "description": "Typed package lifecycle outcome truth.",
            "$ref": "#/components/schemas/AgentPackageOperationStateDto"
          },
          "progressPercent": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Bounded server-reported progress percentage for the typed operation.",
            "format": "int32"
          },
          "stage": {
            "type": "string",
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          },
          "detail": {
            "type": "string",
            "description": "Safe operator detail without credentials, raw paths, SQL, shell text, private keys, or exception data."
          },
          "rollbackAttempted": {
            "type": "boolean",
            "description": "Boolean outcome or capability flag produced by the server-owned boundary."
          },
          "rollbackSucceeded": {
            "type": "boolean",
            "description": "Boolean outcome or capability flag produced by the server-owned boundary."
          },
          "recoveryRequired": {
            "type": "boolean",
            "description": "Boolean outcome or capability flag produced by the server-owned boundary."
          },
          "startedAtUtc": {
            "type": "string",
            "description": "UTC lifecycle timestamp produced and enforced by the Agent.",
            "format": "date-time"
          },
          "completedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "description": "UTC lifecycle timestamp produced and enforced by the Agent.",
            "format": "date-time"
          },
          "correlationId": {
            "type": "string",
            "description": "Safe request correlation identifier for operator diagnostics."
          }
        },
        "description": "Principal-scoped package lifecycle state with activation, health, rollback, and recovery truth."
      },
      "AgentPackageOperationKindDto": {
        "enum": [
          "install",
          "upgrade",
          "uninstall",
          "rollback",
          "repair",
          "health"
        ],
        "description": "Server-owned logical label or profile value; it is not a browser-selected path or endpoint."
      },
      "AgentPackageOperationRequestDto": {
        "required": [
          "previewId",
          "typedConfirmation",
          "idempotencyKey"
        ],
        "type": "object",
        "properties": {
          "previewId": {
            "type": "string",
            "description": "Opaque Agent-issued identifier scoped to the authenticated principal and typed workflow."
          },
          "typedConfirmation": {
            "type": "string",
            "description": "Exact server-owned typed confirmation value required for this workflow boundary."
          },
          "idempotencyKey": {
            "type": "string",
            "description": "Bounded caller-generated key used to prevent duplicate execution of the same typed request."
          },
          "snapshotId": {
            "type": [
              "null",
              "string"
            ],
            "description": "Opaque Agent-issued identifier scoped to the authenticated principal and typed workflow."
          }
        },
        "description": "Typed request for one prior Agent package preview with exact confirmation and idempotency."
      },
      "AgentPackageOperationStateDto": {
        "enum": [
          "notAttempted",
          "preview",
          "accepted",
          "staging",
          "activating",
          "verifying",
          "completed",
          "failed",
          "rollbackSucceeded",
          "rollbackFailed",
          "recoveryRequired",
          "outcomeUnknown"
        ],
        "description": "Typed server-owned state classification; unknown and recovery states are preserved explicitly."
      },
      "AgentPackagePreviewDto": {
        "required": [
          "previewId",
          "operation",
          "ready",
          "verification",
          "manifest",
          "effects",
          "blockers",
          "confirmationPhrase",
          "expiresAtUtc"
        ],
        "type": "object",
        "properties": {
          "previewId": {
            "type": "string",
            "description": "Opaque Agent-issued identifier scoped to the authenticated principal and typed workflow."
          },
          "operation": {
            "description": "Typed package lifecycle operation selected by the server-owned policy.",
            "$ref": "#/components/schemas/AgentPackageOperationKindDto"
          },
          "ready": {
            "type": "boolean",
            "description": "Boolean outcome or capability flag produced by the server-owned boundary."
          },
          "verification": {
            "description": "Typed package signature, checksum, compatibility, and trust verification state.",
            "$ref": "#/components/schemas/AgentPackageVerificationStateDto"
          },
          "manifest": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AgentPackageManifestDto"
              }
            ],
            "description": "Safe verified package manifest projection, when available."
          },
          "effects": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Bounded server-owned logical evidence or precondition labels; raw machine targets are omitted."
          },
          "blockers": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Bounded server-owned logical evidence or precondition labels; raw machine targets are omitted."
          },
          "confirmationPhrase": {
            "type": "string",
            "description": "Exact server-owned typed confirmation value required for this workflow boundary."
          },
          "expiresAtUtc": {
            "type": "string",
            "description": "UTC lifecycle timestamp produced and enforced by the Agent.",
            "format": "date-time"
          }
        },
        "description": "Typed Agent package lifecycle preview with manifest verification and bounded effects/blockers."
      },
      "AgentPackageStatusDto": {
        "required": [
          "installedVersion",
          "previousVersion",
          "verification",
          "state",
          "manifest",
          "detail"
        ],
        "type": "object",
        "properties": {
          "installedVersion": {
            "type": [
              "null",
              "string"
            ],
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          },
          "previousVersion": {
            "type": [
              "null",
              "string"
            ],
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          },
          "verification": {
            "description": "Typed package signature, checksum, compatibility, and trust verification state.",
            "$ref": "#/components/schemas/AgentPackageVerificationStateDto"
          },
          "state": {
            "description": "Typed installed package lifecycle state.",
            "$ref": "#/components/schemas/AgentPackageOperationStateDto"
          },
          "manifest": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AgentPackageManifestDto"
              }
            ],
            "description": "Safe verified package manifest projection, when available."
          },
          "detail": {
            "type": "string",
            "description": "Safe operator detail without credentials, raw paths, SQL, shell text, private keys, or exception data."
          }
        },
        "description": "Safe server-owned Agent package status and truthful verification state."
      },
      "AgentPackageVerificationStateDto": {
        "enum": [
          "unverified",
          "verified",
          "rejected",
          "unknown"
        ],
        "description": "Typed server-owned state classification; unknown and recovery states are preserved explicitly."
      },
      "AgentProblemDetailsDto": {
        "required": [
          "type",
          "title",
          "status"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Problem type identifier for the stable Agent error contract."
          },
          "title": {
            "type": "string",
            "description": "Safe human-readable explanation of why the Agent rejected the request."
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "HTTP status code returned by the Agent.",
            "format": "int32"
          },
          "code": {
            "type": [
              "null",
              "string"
            ],
            "description": "Optional stable machine-readable Agent problem code used by clients; it never carries raw identity or credential data."
          },
          "correlationId": {
            "type": [
              "null",
              "string"
            ],
            "description": "Optional request correlation identifier echoed by the Agent for diagnostics."
          }
        },
        "description": "Safe application/problem+json error contract. Code is a stable problem code and correlationId may identify the request; sensitive identity and machine details are not included."
      },
...
    },
    "securitySchemes": {
      "WindowsNegotiate": {
        "type": "http",
        "description": "Windows Integrated Authentication using Negotiate. The browser/OS supplies credentials with a credentialed cross-origin request; clients must rely on the browser's integrated credential flow.",
        "scheme": "negotiate"
      }
    }
  },
...
}